StepTimer#max should always return max of last period (#1856)
Also introduces the abstract type StepValue, which is used by the different variants like StepDouble and StepLong. This adds the new type StepLongMax, which is used as the implementation to change the behavior of StepTimer's max value to return the max of the previous step.

Resolves #1796

+++ micrometer-core/src/main/java/io/micrometer/core/instrument/step/StepDistributionSummary.java
@@ -88,7 +88,7 @@ public class StepDistributionSummary extends AbstractDistributionSummary {
 
     @Override
     public long count() {
-        return (long) count.poll();
+        return count.poll();
     }
 
     @Override
